requirements.txt
fastapi
uvicorn
python-dotenv
json_repair
agno
groq
